requirements.txt
asgiref==3.7.2
Django~=4.2.6
sqlparse==0.4.4
tzdata==2023.3
ipython
psycopg2-binary
django-filter
pillow
djangorestframework-simplejwt

python-dotenv~=1.0.0
APScheduler~=3.10.4
djangorestframework~=3.14.0